move value serializer to its own file
