+++ src/ser/raw/document_serializer.rs
@@ -1,22 +1,26 @@
+use serde::Serialize;
+use serde::ser::Impossible;
+
+use crate::ser::write_cstring;
 use crate::ser::write_i32;
 use crate::ser::Result;
 use crate::ser::Error;
 
 use super::Serializer;
 
-pub(super) struct DocumentSerializationResult<'a> {
-    pub(super) length: i32,
-    pub(super) root_serializer: &'a mut Serializer,
+pub(crate) struct DocumentSerializationResult<'a> {
+    pub(crate) length: i32,
+    pub(crate) root_serializer: &'a mut Serializer,
 }
 
-pub(super) struct DocumentSerializer<'a> {
+pub(crate) struct DocumentSerializer<'a> {
     root_serializer: &'a mut Serializer,
     num_keys_serialized: usize,
     start: usize,
 }
 
 impl<'a> DocumentSerializer<'a> {
-    pub(super) fn start(rs: &'a mut Serializer) -> crate::ser::Result<Self> {
+    pub(crate) fn start(rs: &'a mut Serializer) -> crate::ser::Result<Self> {
         let start = rs.bytes.len();
         write_i32(&mut rs.bytes, 0)?;
         Ok(Self {
@@ -41,7 +45,7 @@ impl<'a> DocumentSerializer<'a> {
         Ok(())
     }
 
-    pub(super) fn end_doc(self) -> crate::ser::Result<DocumentSerializationResult<'a>> {
+    pub(crate) fn end_doc(self) -> crate::ser::Result<DocumentSerializationResult<'a>> {
         self.root_serializer.bytes.push(0);
         let length = (self.root_serializer.bytes.len() - self.start) as i32;
         self.root_serializer.replace_i32(self.start, length);
@@ -144,5 +148,171 @@ impl<'a> serde::ser::SerializeTupleStruct for DocumentSerializer<'a> {
 
     fn end(self) -> Result<Self::Ok> {
         self.end_doc().map(|_| ())
+    }
+}
+
+struct KeySerializer<'a> {
+    root_serializer: &'a mut Serializer,
+}
+
+impl<'a> serde::Serializer for KeySerializer<'a> {
+    type Ok = ();
+
+    type Error = Error;
+
+    type SerializeSeq = Impossible<(), Error>;
+    type SerializeTuple = Impossible<(), Error>;
+    type SerializeTupleStruct = Impossible<(), Error>;
+    type SerializeTupleVariant = Impossible<(), Error>;
+    type SerializeMap = Impossible<(), Error>;
+    type SerializeStruct = Impossible<(), Error>;
+    type SerializeStructVariant = Impossible<(), Error>;
+
+    fn serialize_bool(self, v: bool) -> Result<Self::Ok> {
+        todo!()
+    }
+
+    fn serialize_i8(self, v: i8) -> Result<Self::Ok> {
+        todo!()
+    }
+
+    fn serialize_i16(self, v: i16) -> Result<Self::Ok> {
+        todo!()
+    }
+
+    fn serialize_i32(self, v: i32) -> Result<Self::Ok> {
+        todo!()
+    }
+
+    fn serialize_i64(self, v: i64) -> Result<Self::Ok> {
+        todo!()
+    }
+
+    fn serialize_u8(self, v: u8) -> Result<Self::Ok> {
+        todo!()
+    }
+
+    fn serialize_u16(self, v: u16) -> Result<Self::Ok> {
+        todo!()
+    }
+
+    fn serialize_u32(self, v: u32) -> Result<Self::Ok> {
+        todo!()
+    }
+
+    fn serialize_u64(self, v: u64) -> Result<Self::Ok> {
+        todo!()
+    }
+
+    fn serialize_f32(self, v: f32) -> Result<Self::Ok> {
+        todo!()
+    }
+
+    fn serialize_f64(self, v: f64) -> Result<Self::Ok> {
+        todo!()
+    }
+
+    fn serialize_char(self, v: char) -> Result<Self::Ok> {
+        todo!()
+    }
+
+    fn serialize_str(self, v: &str) -> Result<Self::Ok> {
+        write_cstring(&mut self.root_serializer.bytes, v)
+    }
+
+    fn serialize_bytes(self, v: &[u8]) -> Result<Self::Ok> {
+        todo!()
+    }
+
+    fn serialize_none(self) -> Result<Self::Ok> {
+        todo!()
+    }
+
+    fn serialize_some<T: ?Sized>(self, value: &T) -> Result<Self::Ok>
+    where
+        T: Serialize,
+    {
+        todo!()
+    }
+
+    fn serialize_unit(self) -> Result<Self::Ok> {
+        todo!()
+    }
+
+    fn serialize_unit_struct(self, name: &'static str) -> Result<Self::Ok> {
+        todo!()
+    }
+
+    fn serialize_unit_variant(
+        self,
+        name: &'static str,
+        variant_index: u32,
+        variant: &'static str,
+    ) -> Result<Self::Ok> {
+        todo!()
+    }
+
+    fn serialize_newtype_struct<T: ?Sized>(self, name: &'static str, value: &T) -> Result<Self::Ok>
+    where
+        T: Serialize,
+    {
+        todo!()
+    }
+
+    fn serialize_newtype_variant<T: ?Sized>(
+        self,
+        name: &'static str,
+        variant_index: u32,
+        variant: &'static str,
+        value: &T,
+    ) -> Result<Self::Ok>
+    where
+        T: Serialize,
+    {
+        todo!()
+    }
+
+    fn serialize_seq(self, len: Option<usize>) -> Result<Self::SerializeSeq> {
+        todo!()
+    }
+
+    fn serialize_tuple(self, len: usize) -> Result<Self::SerializeTuple> {
+        todo!()
+    }
+
+    fn serialize_tuple_struct(
+        self,
+        name: &'static str,
+        len: usize,
+    ) -> Result<Self::SerializeTupleStruct> {
+        todo!()
+    }
+
+    fn serialize_tuple_variant(
+        self,
+        name: &'static str,
+        variant_index: u32,
+        variant: &'static str,
+        len: usize,
+    ) -> Result<Self::SerializeTupleVariant> {
+        todo!()
+    }
+
+    fn serialize_map(self, len: Option<usize>) -> Result<Self::SerializeMap> {
+        todo!()
+    }
+
+    fn serialize_struct(self, name: &'static str, len: usize) -> Result<Self::SerializeStruct> {
+        todo!()
+    }
+
+    fn serialize_struct_variant(
+        self,
+        name: &'static str,
+        variant_index: u32,
+        variant: &'static str,
+        len: usize,
+    ) -> Result<Self::SerializeStructVariant> {
+        todo!()
     }
 }
